Separated out users component
